requirements.txt
# Core dependencies
pynput==1.7.6
pyperclip==1.8.2
requests==2.31.0
python-dotenv==1.0.1

# System tray icon dependencies
pillow==10.2.0
pystray==0.19.5

# Optional dependencies for better compatibility
six>=1.16.0
pywin32>=306; sys_platform == 'win32'
